+++ src/lib/services/evm/index.ts
@@ -11,6 +11,7 @@ import { isHexWalletAddress } from "lib/utils";
 import {
   getEvmCodesByAddress,
   getEvmContractInfoSequencer,
+  getEvmDenomByAddressLcd,
   getEvmParams,
 } from "./lcd";
 
@@ -29,16 +30,16 @@ export const useEvmParams = () => {
   );
 };
 
-export const useEvmCodesByAddress = (address: HexAddr20) => {
+export const useEvmCodesByAddress = (address: HexAddr20, enabled = true) => {
   const lcdEndpoint = useLcdEndpoint();
 
   return useQuery(
-    [CELATONE_QUERY_KEYS.EVM_CODES_BY_ADDRESS_LCD, lcdEndpoint],
+    [CELATONE_QUERY_KEYS.EVM_CODES_BY_ADDRESS_LCD, lcdEndpoint, address],
     async () => getEvmCodesByAddress(lcdEndpoint, address),
     {
       retry: 1,
       refetchOnWindowFocus: false,
-      enabled: address && isHexWalletAddress(address),
+      enabled: enabled && address && isHexWalletAddress(address),
     }
   );
 };
@@ -50,8 +51,27 @@ export const useEvmContractInfoSequencer = (address: HexAddr20) => {
   } = useCurrentChain();
 
   return useQuery(
-    [CELATONE_QUERY_KEYS.EVM_CONTRACT_INFO_SEQUENCER, lcdEndpoint, prefix],
+    [
+      CELATONE_QUERY_KEYS.EVM_CONTRACT_INFO_SEQUENCER,
+      lcdEndpoint,
+      prefix,
+      address,
+    ],
     async () => getEvmContractInfoSequencer(lcdEndpoint, prefix, address),
+    {
+      retry: 1,
+      refetchOnWindowFocus: false,
+      enabled: address && isHexWalletAddress(address),
+    }
+  );
+};
+
+export const useEvmDenomByAddressLcd = (address: HexAddr20) => {
+  const lcdEndpoint = useLcdEndpoint();
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.EVM_DENOM_BY_ADDRESS_LCD, lcdEndpoint, address],
+    async () => getEvmDenomByAddressLcd(lcdEndpoint, address),
     {
       retry: 1,
       refetchOnWindowFocus: false,