+++ src/lib/services/tx/index.ts
@@ -45,6 +45,7 @@ import {
   getTxsCountByAddress,
 } from "./api";
 import {
+  getCosmosTxHashByEvmTxHash,
   getEvmTxHashByCosmosTxHash,
   getEvmTxHashesByCosmosTxHashes,
   getTxDataJsonRpc,
@@ -77,7 +78,7 @@ export const useTxData = (
 
   const queryFn = useCallback(
     async (hash: Option<string>) => {
-      if (!hash) throw new Error("CELATONE_QUERY_KEYS.TX_DATA is undefined");
+      if (!hash) throw new Error("hash is undefined (useTxData)");
 
       const txData = isFullTier
         ? await getTxData(endpoint, hash)
@@ -163,7 +164,7 @@ export const useTxsByAddress = (
       isMove,
     ],
     async () => {
-      if (!address) throw new Error("No user address");
+      if (!address) throw new Error("address is undefined (useTxsByAddress)");
       return getTxsByAddress(
         endpoint,
         address,
@@ -591,7 +592,7 @@ export const useEvmTxHashesByCosmosTxHashes = (
   );
 };
 
-export const useTxDataJsonRpc = (evmTxHash: string) => {
+export const useTxDataJsonRpc = (evmTxHash: string, enabled = true) => {
   const evm = useEvmConfig({ shouldRedirect: false });
 
   return useQuery(
@@ -605,6 +606,29 @@ export const useTxDataJsonRpc = (evmTxHash: string) => {
         throw new Error("EVM is not enabled (useTxDataJsonRpc)");
 
       return getTxDataJsonRpc(evm.jsonRpc, evmTxHash);
+    },
+    {
+      retry: false,
+      refetchOnWindowFocus: false,
+      enabled: enabled && evm.enabled && !!evm.jsonRpc,
+    }
+  );
+};
+
+export const useCosmosTxHashByEvmTxHash = (evmTxHash: string) => {
+  const evm = useEvmConfig({ shouldRedirect: false });
+
+  return useQuery(
+    [
+      CELATONE_QUERY_KEYS.COSMOS_TX_HASH_BY_EVM_TX_HASH,
+      evm.enabled && evm.jsonRpc,
+      evmTxHash,
+    ],
+    async () => {
+      if (!evm.enabled)
+        throw new Error("EVM is not enabled (useCosmosTxHashByEvmTxHash)");
+
+      return getCosmosTxHashByEvmTxHash(evm.jsonRpc, evmTxHash);
     },
     {
       retry: false,