+++ src/lib/utils/number.ts
@@ -1,3 +1,5 @@
+import { big } from "lib/types";
+
 export const isPositiveInt = (str: string): boolean =>
   /^\+?([1-9]\d*)$/.test(str);
 
@@ -5,3 +7,10 @@ export const isNumeric = (str: string): boolean =>
   /^-?\d+(?:\.\d+)?$/.test(str);
 
 export const numberToHex = (num: number): string => `0x${num.toString(16)}`;
+
+export const hexToBig = (hex: string) => {
+  if (!hex.startsWith("0x")) {
+    return big(BigInt(`0x${hex}`).toString());
+  }
+  return big(BigInt(hex).toString());
+};